+++ server/src/models/data/index.ts
@@ -1,1 +1,2 @@
 export * from './language-levels.data';
+export * from './available-languages.data';
